* commit '63264204dc372093e2af1b22e63b71ac5b8a98ce':
  bump minor version
  change ids to long
  support refreshing of oauth tokens

+++ api/src/main/java/com/ripstech/apiconnector2/entity/send/application/scan/issue/ReviewSend.java
@@ -5,16 +5,16 @@ import com.fasterxml.jackson.annotation.JsonRootName;
 @JsonRootName("review")
 public class ReviewSend {
 
-	private Integer type;
+	private Long type;
 	private String source;
 
-	public static ReviewSend createPost(int type) {
+	public static ReviewSend createPost(long type) {
 		return new ReviewSend().setType(type);
 	}
 
 	private ReviewSend() {}
 
-	public Integer getType() {
+	public Long getType() {
 		return this.type;
 	}
 
@@ -22,7 +22,7 @@ public class ReviewSend {
 		return this.source;
 	}
 
-	public ReviewSend setType(Integer type) {
+	public ReviewSend setType(Long type) {
 		this.type = type;
 		return this;
 	}
